+++ espflash/src/target/efuse/esp32c3.rs
@@ -2,15 +2,60 @@
 //!
 //! This file was automatically generated, please do not edit it manually!
 //!
-//! Generated: 2025-11-19 12:31
+//! Generated: 2025-11-26 21:42
 //! Version:   4622cf9245401eca0eb1df8122449a6d
 
 #![allow(unused)]
 
-use super::EfuseField;
+use super::{EfuseBlock, EfuseField};
 
-/// Total size in bytes of each block
-pub(crate) const BLOCK_SIZES: &[u32] = &[24, 24, 32, 32, 32, 32, 32, 32, 32, 32, 32];
+/// All eFuse blocks available on this device.
+pub(crate) const BLOCKS: &[EfuseBlock] = &[
+    EfuseBlock {
+        length: 6u8,
+        read_address: 1610647596u32,
+    },
+    EfuseBlock {
+        length: 6u8,
+        read_address: 1610647620u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647644u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647676u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647708u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647740u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647772u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647804u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647836u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647868u32,
+    },
+    EfuseBlock {
+        length: 8u8,
+        read_address: 1610647900u32,
+    },
+];
 
 /// Disable programming of individual eFuses
 pub const WR_DIS: EfuseField = EfuseField::new(0, 0, 0, 32);