+++ CHANGELOG.md
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Corrected eFuse BLOCK0 definitions for ESP32-C2, ESP32-C3, and ESP32-S3 (#961)
+- Corrected eFuse block address calculations. (#971)
 - Fixed Secure Download Mode detection on ESP32-P4 (#972)
 - Several fixes in `read_efuse` (#969)
 - Fixed a problem in detecting the app-descriptor for a project if `strip = true` is used (#975)